Replace AircloakAuthenticationHandler.
AircloakAuthenticationHandler inherited from DelegatingHandler which
meant it need to be registered in a chain of handlers. This proved
tricky to achieve cleanly for the Api App and test environment.

The new IAircloakAuthenticationProvider doesn't derive from anything
and is consumed as a dependency in the JsonApiClient constructor. This
simplifies the implementation and makes usage more flexible
(ie. choice between using DI vs. explicit construction).

The existing *AuthHandlers have been migrated to the *AuthProvider
interface and service registration methods and testing have also been
adapted.

I also bumped Microsoft.AspNetCore.Mvc.Testing to the latest version.