Resolve an issue with resolution of publishers
This became apparent when adding the component test
`TestDistinctValuesMetricContainsRemainder`.
It appears that PublisherComponents are registered against the interface
and so therefore if you request the concrete class through DI, you get an
instance with transient lifetime instead of scoped (which is desired).
This has been fixed by adding an extension method to request the
appropriate PublisherComponent through its interface using Lamar's
name resolution convention.

(If that sounds cryptic, hopefully the code change in `LamarExtensions` is
somewhat clearer.