Consider distinct values for min/max estimation.
In some cases min and max values aren't returned from the min and max sql
functions (returns null instead). This change adds a check for this and
uses the distinct values as backup to try to estimate a min and max.

This change also adds the distinct values metric to all result sets,
irrespective of whether the column is considered categorical or not.

I also replaced the generic `ILogger` with explicit `ILogger<T>` for
`ExploreComponent`s since the ILogger was not being correcly
injected by Lamar.